+++ lib/aws-sdk-rails.rb
@@ -1,4 +1,6 @@
-require_relative 'aws/rails/mailer'
+require_relative 'aws/rails/ses_mailer'
+require_relative 'aws/rails/pinpoint_mailer'
+require_relative 'aws/rails/aws_sdk'
 
 module Aws
   module Rails
@@ -7,7 +9,10 @@ module Aws
     class Railtie < ::Rails::Railtie
       initializer "aws-sdk-rails.initialize", before: :load_config_initializers do |app|
         # Initialization Actions
-        Aws::Rails.add_action_mailer_delivery_method
+        [:ses_mailer, :pinpoint_mailer, :aws_sdk].each do |name|
+          Aws::Rails.add_action_mailer_delivery_method name
+        end
+
         Aws::Rails.log_to_rails_logger
       end
     end
@@ -20,9 +25,9 @@ module Aws
     #   register.
     # @param [Hash] options The options you wish to pass on to the
     #   Aws::SES::Client initialization method.
-    def self.add_action_mailer_delivery_method(name = :aws_sdk, options = {})
+    def self.add_action_mailer_delivery_method(name = :ses_mailer, options = {})
       ActiveSupport.on_load(:action_mailer) do
-        self.add_delivery_method(name, Aws::Rails::Mailer, options)
+        self.add_delivery_method(name, "Aws::Rails::#{name.to_s.classify}".constantize, options)
       end
     end